Rename gcobj to object_g to align with other types
Align the convention used for GC-safe pointers:
they all end in `_g`.

+++ src/tests.cc
@@ -692,9 +692,9 @@ tests &tests::itest(long long value)
 // ----------------------------------------------------------------------------
 {
     if (value < 0)
-        return itest(-value, CHS);
+        return itest((unsigned long long) -value, CHS);
     else
-        return itest(value);
+        return itest((unsigned long long) value);
 }
